Add refrigeration parts to the cycle builder
This update introduces a significant change in the Thercy cycle building process: it now supports evaporator and heat exchanger. In addition, the rankine cycle build example was updated, and a new refrigeration build example was added. It's now possible to create more specific and complex cycles, enhancing the overall flexibility of the cycle builder system. Minor refactoring was also done..